Moved common test logic to a shared class.
